// gpc-front-end/src/app/shared/enums/Group.enum.ts
export enum Group {
    Admin = 1,
    OperatoreSportello = 2, 
    DirettoreMunicipio = 3, 
    IstruttoreMunicipio = 4, 
    PoliziaLocale = 5, 
    UfficioTecnicoDecentrato = 6,
    RipartizioneUrbanistica = 7,
    AmministratorePassiCarrabili = 8,
    RipartizioneTributi = 11,
    RipartizioneRagioneria = 12,
    Concessionario = 13,
    PoliziaLocaleControlliSulTerritorio = 14
}